refactor(discord): extract task response authority
