+++ src/services/discord/session_relay_sink.rs
@@ -45,13 +45,8 @@ use self::idle_jsonl::{
     idle_jsonl_should_retry_without_dedup_shared, idle_relay_range_action,
     prune_idle_jsonl_session_state, read_jsonl_range,
 };
-use self::task_notification_context::{
-    answer_reference, commit_response_fence, ensure_card_and_route, merge_task_notification_kind,
-};
-use super::task_notification_delivery::{
-    ResponseDeliveryClaim, ResponseDeliveryClaimOutcome, renew_task_response_delivery,
-    task_response_delivery_heartbeat,
-};
+use self::task_notification_context::{ensure_card_and_route, merge_task_notification_kind};
+use super::task_notification_delivery::{ResponseDeliveryClaim, ResponseDeliveryClaimOutcome};
 
 static SESSION_BOUND_DISCORD_DELIVERY_ENABLED: AtomicBool = AtomicBool::new(false);
 const IDLE_JSONL_RELAY_POLL_INTERVAL: Duration = Duration::from_millis(500);
@@ -1128,98 +1123,19 @@ impl SessionBoundDiscordRelaySink {
                 }
             }
         } else {
-            let prompt_anchor = relay_format::ssh_direct_prompt_anchor_for_response(
-                &provider,
-                &delivery.session_name,
-                channel_id,
-            );
-            let prompt_anchor_reference =
-                answer_reference(channel, task_card_message_id, prompt_anchor);
-            if let Some(task_card_message_id) = task_card_message_id {
-                if let Some(claim) = task_response_claim.as_ref() {
-                    renew_task_response_delivery(shared.pg_pool.as_ref(), claim)
-                        .await
-                        .map_err(|error| {
-                            RelaySinkError::Transient(format!(
-                                "task response claim was lost before send: {error}"
-                            ))
-                        })?;
-                }
-                let heartbeat = task_response_delivery_heartbeat(
-                    shared.pg_pool.as_ref(),
-                    task_response_claim.as_ref(),
-                );
-                let send_result =
-                    formatting::send_long_message_raw_with_required_reference_rollback(
-                        &http,
-                        channel,
-                        task_card_message_id,
-                        &relay_text,
-                        &shared,
-                        (channel, task_card_message_id),
-                    )
-                    .await;
-                heartbeat.stop();
-                send_result.map_err(|error| RelaySinkError::Transient(error.to_string()))?;
-            } else {
-                formatting::send_long_message_raw_with_reference(
-                    &http,
-                    channel,
-                    &relay_text,
-                    &shared,
-                    prompt_anchor_reference,
-                )
-                .await
-                .map_err(|error| RelaySinkError::Transient(error.to_string()))?;
-            }
-            if let Some(prompt_anchor) = prompt_anchor {
-                relay_format::clear_ssh_direct_prompt_anchor(
-                    &provider,
-                    &delivery.session_name,
-                    prompt_anchor,
-                );
-            }
-            self.delivered_total.fetch_add(1, Ordering::AcqRel);
-            // #3041 P1-4 (§4-④): lease released by the RAII guard on exit.
-            tracing::info!(
-                provider = provider.as_str(),
-                channel_id,
-                tmux_session = %delivery.session_name,
-                turn_id = trace.turn_id().unwrap_or(""),
-                dispatch_id = trace.dispatch_id().unwrap_or(""),
-                session_key = trace.session_key().unwrap_or(""),
-                relay_owner = trace.relay_owner(),
-                runtime_kind = trace.runtime_kind(),
-                prompt_anchor_message_id = prompt_anchor_reference
-                    .map(|(_, message_id)| message_id.get()),
-                chars = relay_text.chars().count(),
-                "session-bound relay sink delivered terminal response via new message"
-            );
-            crate::services::observability::emit_relay_delivery(
-                provider.as_str(),
-                channel_id,
-                trace.dispatch_id(),
-                trace.session_key(),
-                trace.turn_id(),
-                prompt_anchor_reference.map(|(_, message_id)| message_id.get()),
-                "session_relay_sink",
-                "post",
-                None,
-                None,
-                true,
-                Some("new message"),
-            );
-            // #3041 P1-3 (Part a, B1, codex issue 3): commit fence — post-POST fresh re-check before advance.
-            self.advance_after_confirmed_post(
+            self.deliver_new_message_with_task_authority(
+                &http,
                 &shared,
                 &provider,
                 channel_id,
-                &delivery.session_name,
                 &delivery,
+                &relay_text,
+                task_card_message_id,
+                task_response_claim.as_ref(),
+                &trace,
                 sink_lease_guard.as_ref(),
-            );
-            commit_response_fence(&shared, &delivery, task_response_claim.as_ref()).await;
-            Ok(SessionRelayDeliveryOutcome::Delivered)
+            )
+            .await
         }
     }
 }
